README.md run game jar locally

+++ src/main/java/com/core/providers/StorageProvider.java
@@ -6,6 +6,8 @@ import com.opencsv.CSVReader;
 import com.opencsv.CSVWriter;
 
 import java.io.*;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -15,11 +17,9 @@ public class StorageProvider implements Path {
 
     public static List<User> getStorage(){
        Map<String, String> dataMap;
-       try {
-        dataMap = readCsvToHashMap();
-       } catch (IOException e) {
-           throw new RuntimeException(e);
-       }
+
+       dataMap = readCsvToHashMap();
+
        return dataMap.entrySet().stream()
                 .map(entry -> new User(entry.getKey(), Integer.parseInt(entry.getValue())))
                 .collect(Collectors.toList());
@@ -36,39 +36,69 @@ public class StorageProvider implements Path {
         }
     }
 
-    private static Map<String, String> readCsvToHashMap() throws IOException {
+
+    private static Map<String, String> readCsvToHashMap()  {
         Map<String, String> map = new HashMap<>();
-        try (CSVReader reader = new CSVReader(new FileReader(Path.STORAGE_PATH))) {
-            String[] header = reader.readNext();
-            if (header == null) {
-                return map;
-            }
-            String[] row;
-            boolean hasData = false;
+        try {
+            File directory = new File("src/main/resources");
+            File file = new File(Path.STORAGE_PATH);
 
-            while ((row = reader.readNext()) != null) {
-                if (row.length < 2) {
-                    System.err.println("Invalid row: " + String.join(",", row));
-                    continue;
+            if (!directory.exists()) {
+                directory.mkdirs();
+            }
+
+            if(!file.exists()){
+                file.createNewFile();
+                try (CSVWriter writer = new CSVWriter(new FileWriter(Path.STORAGE_PATH))) {
+                    writer.writeNext(new String[]{"Name", "Score"});
                 }
-                map.put(row[0], row[1]);
-                hasData = true;
             }
 
-            if (!hasData) {
-                System.out.println("The storage.csv contains no data.");
+            try (CSVReader reader = new CSVReader(new FileReader(Path.STORAGE_PATH))) {
+                String[] header = reader.readNext();
+                if (header == null) {
+                    return map;
+                }
+                String[] row;
+                boolean hasData = false;
+
+                while ((row = reader.readNext()) != null) {
+                    if (row.length < 2) {
+                        System.err.println("Invalid row: " + String.join(",", row));
+                        continue;
+                    }
+                    map.put(row[0], row[1]);
+                    hasData = true;
+                }
+
+                if (!hasData) {
+                    System.out.println("The storage.csv contains no data.");
+                }
             }
+            return map;
+        }catch (RuntimeException | IOException e){
+            System.out.println(" here readCsvToHashMap" );
         }
         return map;
     }
 
     private static void writeMapToCsv(Map<String, String> map) throws IOException {
-        try (CSVWriter writer = new CSVWriter(new FileWriter(Path.STORAGE_PATH))) {
-            writer.writeNext(new String[]{"Name", "Score"});
+        try {
 
-            for (Map.Entry<String, String> entry : map.entrySet()) {
-                writer.writeNext(new String[]{entry.getKey(), entry.getValue()});
+            File file = new File(Path.STORAGE_PATH);
+            if(!file.exists()){
+                file.createNewFile();
             }
+
+            try (CSVWriter writer = new CSVWriter(new FileWriter(Path.STORAGE_PATH))) {
+                writer.writeNext(new String[]{"Name", "Score"});
+
+                for (Map.Entry<String, String> entry : map.entrySet()) {
+                    writer.writeNext(new String[]{entry.getKey(), entry.getValue()});
+                }
+            }
+        }catch (RuntimeException e){
+            System.out.println("here writeMapToCsv");
         }
     }
 }
